web/requirements.txt
fastapi
uvicorn
asyncio-mqtt
aiofiles
pyyaml
